The Start of the End -
Got Idea to simply display what both a player interface has, and what a DM has. Other wise we need to implement more data model's to contain user accounts and campaigns. We don't have time for that. If we can flush out a simple home screen, with a good user interface for BOTH player and DM Interfaces, then we should be good for this Tuesday and Wednesday.